requirements.txt
geopy
folium
geopandas
matplotlib
contextily
pyproj
numpy
pandas
langchain==0.2.16
sqlalchemy
psycopg2-binary
tabulate
shapely
geoalchemy2
python-dotenv
langchain-openai==0.1.23
unidecode
gradio